+++ src/main/java/de/samply/email/EmailContext.java
@@ -1,6 +1,7 @@
 package de.samply.email;
 
 import de.samply.app.ProjectManagerConst;
+import de.samply.utils.Base64Utils;
 import jakarta.annotation.PostConstruct;
 import lombok.Data;
 import lombok.extern.slf4j.Slf4j;
@@ -22,7 +23,13 @@ public class EmailContext {
     @PostConstruct
     public void init() {
         log.info("Email Context initialized:");
-        context.entrySet().stream().forEach(keyValue -> log.info("\t-\t{}: {}", keyValue.getKey(), keyValue.getValue()));
+        Map<String, String> tempContext = Map.copyOf(context);
+        context.entrySet().stream().forEach(keyValue -> {
+            log.info("\t-\t{}: {}", keyValue.getKey(), keyValue.getValue());
+            Base64Utils.decodeIfNecessary(keyValue.getValue()).ifPresent(value -> tempContext.put(keyValue.getKey(), value));
+        });
+        context = tempContext;
     }
+
 
 }